requirements.txt
python-dotenv==1.0.0
ics==0.7.2
DateTime==5.2